refact:Renomeando Class1 para EventManager
